// lib/netlify/schema.ts
export interface components {
    schemas: {
// ...
        site: {
            id?: string
            state?: string
            plan?: string
            name?: string
            custom_domain?: string
            domain_aliases?: string[]
            branch_deploy_custom_domain?: string
            deploy_preview_custom_domain?: string
            password?: string
            notification_email?: string
            url?: string
            ssl_url?: string
            admin_url?: string
            screenshot_url?: string
            /** Format: dateTime */
            created_at?: string
            /** Format: dateTime */
            updated_at?: string
            user_id?: string
            session_id?: string
            ssl?: boolean
            force_ssl?: boolean
            managed_dns?: boolean
            deploy_url?: string
            published_deploy?: components['schemas']['deploy']
            account_id?: string
            account_name?: string
            account_slug?: string
            git_provider?: string
            deploy_hook?: string
            capabilities?: {
                [key: string]: Record<string, never>
            }
            processing_settings?: {
                html?: {
                    pretty_urls?: boolean
                }
                ignore_html_forms?: boolean
            }
            build_settings?: components['schemas']['repoInfo']
            id_domain?: string
            default_hooks_data?: {
                access_token?: string
            }
            build_image?: string
            prerender?: string
            functions_region?: string
        }
// ...
    }
// ...
}